# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.BasicResponseDTO:
    properties:
      data: {}
      status:
        type: integer
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: chat
        type: string
    type: object
...
  payment.Response:
    properties:
      ad_id:
        type: integer
      confirmation_link:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      status:
        type: string
      tariff:
        $ref: '#/definitions/tariff.Tariff'
      transaction_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  payment.StatusChangedResponse:
    properties:
      status:
        type: string
    type: object
  tariff.Tariff:
    properties:
      createdAt:
        type: string
      currency:
        type: string
      description:
        type: string
      durationMin:
        type: integer
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      price:
        type: number
      updatedAt:
        type: string
    type: object
...
host: http://localhost:8989
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Ads Service API
  version: "1.0"
paths:
...
  /api/v1/payments:
    post:
      consumes:
      - application/json
      description: Регистрация платежа и генерация ссылки
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/payment.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание платежа
      tags:
      - Payments
  /api/v1/payments/webhook:
    post:
      consumes:
      - application/json
      description: Обработка вебхука и смена статуса платежа
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.BasicResponseDTO'
            - properties:
                data:
                  $ref: '#/definitions/payment.StatusChangedResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Нотификация платежа
      tags:
      - Payments
swagger: "2.0"